Moved old tools into new tools_legacy directory
