+++ bfasst/flows_legacy/gather_impl_data.py
@@ -5,13 +5,13 @@
 import shutil
 
 from bfasst.flows_legacy.flow import Flow
-from bfasst.impl.ic2 import Ic2ImplementationTool
-from bfasst.opt.ic2_lse import Ic2LseOptTool
-from bfasst.opt.ic2_synplify import Ic2SynplifyOptTool
-from bfasst.reverse_bit.icestorm import IcestormReverseBitTool
-from bfasst.synth.ic2_lse import Ic2LseSynthesisTool
-from bfasst.synth.ic2_synplify import Ic2SynplifySynthesisTool
-from bfasst.synth.yosys import YosysTechSynthTool
+from bfasst.tools_legacy.impl.ic2 import Ic2ImplementationTool
+from bfasst.tools_legacy.opt.ic2_lse import Ic2LseOptTool
+from bfasst.tools_legacy.opt.ic2_synplify import Ic2SynplifyOptTool
+from bfasst.tools_legacy.reverse_bit.icestorm import IcestormReverseBitTool
+from bfasst.tools_legacy.synth.ic2_lse import Ic2LseSynthesisTool
+from bfasst.tools_legacy.synth.ic2_synplify import Ic2SynplifySynthesisTool
+from bfasst.tools_legacy.synth.yosys import YosysTechSynthTool
 from bfasst.types import ToolType
 from bfasst.job import Job
 from bfasst.flows_legacy.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
