+++ bfasst/flows_legacy/ic2_lse_conformal.py
@@ -6,7 +6,7 @@ from bfasst.flows_legacy.flow import Flow
 from bfasst.flows_legacy.sub_flows.conformal import Conformal
 from bfasst.flows_legacy.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
 from bfasst.job import Job
-from bfasst.synth.ic2_lse import Ic2LseSynthesisTool
+from bfasst.tools_legacy.synth.ic2_lse import Ic2LseSynthesisTool
 from bfasst.types import ToolType
