+++ bfasst/flows_legacy/sub_flows/onespin_cmp.py
@@ -2,7 +2,7 @@
 
 # pylint: disable=duplicate-code
 
-from bfasst.compare.onespin import OneSpinCompareTool
+from bfasst.tools_legacy.compare.onespin import OneSpinCompareTool
 from bfasst.flows_legacy.flow import Flow
 from bfasst.job import Job
 from bfasst.types import ToolType
