+++ bfasst/flows_legacy/yosys_tech_lse_conformal.py
@@ -5,11 +5,11 @@
 from bfasst.flows_legacy.sub_flows.conformal import Conformal
 from bfasst.flows_legacy.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
 from bfasst.flows_legacy.flow import Flow
-from bfasst.impl.ic2 import Ic2ImplementationTool
+from bfasst.tools_legacy.impl.ic2 import Ic2ImplementationTool
 from bfasst.job import Job
-from bfasst.opt.ic2_lse import Ic2LseOptTool
-from bfasst.reverse_bit.icestorm import IcestormReverseBitTool
-from bfasst.synth.yosys import YosysTechSynthTool
+from bfasst.tools_legacy.opt.ic2_lse import Ic2LseOptTool
+from bfasst.tools_legacy.reverse_bit.icestorm import IcestormReverseBitTool
+from bfasst.tools_legacy.synth.yosys import YosysTechSynthTool
 from bfasst.types import ToolType
