+++ bfasst/flows_legacy/yosys_tech_synplify_onespin.py
@@ -6,8 +6,8 @@ from bfasst.flows_legacy.flow import Flow
 from bfasst.flows_legacy.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
 from bfasst.flows_legacy.sub_flows.onespin_cmp import OnespinCmp
 from bfasst.job import Job
-from bfasst.opt.ic2_synplify import Ic2SynplifyOptTool
-from bfasst.synth.yosys import YosysTechSynthTool
+from bfasst.tools_legacy.opt.ic2_synplify import Ic2SynplifyOptTool
+from bfasst.tools_legacy.synth.yosys import YosysTechSynthTool
 from bfasst.types import ToolType
