+++ bfasst/tools_legacy/netlist_mapping/structural/automated_block_mapping.py
@@ -1,6 +1,6 @@
 """Algorithm used to mapped netlists based on their structure"""
 
-from bfasst.netlist_mapping.structural.updating_reversed_netlist import (
+from bfasst.tools_legacy.netlist_mapping.structural.updating_reversed_netlist import (
     update_wires_in_reversed_netlist,
 )
