+++ scripts/run_design.py
@@ -9,7 +9,7 @@ import traceback
 from bfasst.design import Design
 from bfasst import paths
 from bfasst.flows_legacy.flow import get_flow, get_flows
-from bfasst.tool import BfasstException
+from bfasst.tools_legacy.tool import BfasstException
 from bfasst.types import ToolType
